+++ services/base_service.py
@@ -1,6 +1,4 @@
-import csv
-import json
-from bson import ObjectId
+
 
 class base_service:
     def __init__(self, repo):
@@ -10,35 +8,12 @@ class base_service:
     def get_all(self):
         print('in service in get method')
         return self.repo.get()
-
     def get_by_id(self, object_id):
         return self.repo.get_by_id(object_id)
-
-    def create(self, path):
-        result = []
-        try:
-            with open(path['csv_file_path'], 'r') as file:
-                reader = csv.DictReader(file)
-                for row in reader:
-                    tender = {
-                            "tender_id": ObjectId(row['tender_id']),
-                            "category": row['category'],
-                            "tender_name": row['tender_name'],
-                            "date": row['date'],
-                            "details": {
-                              "participants": json.loads(row['details__participants']),
-                              "disqualified_participants": json.loads(row['details__disqualified_participants']),
-                              "committee_member": row['details__committee_member']
-                            }
-                    }
-                    result.append(tender)
-            return self.repo.insert(result)
-        except Exception as e:
-            return e
-
+    def create(self, data):
+        return self.repo.insert(data)
     def update(self, object_id, data):
         return self.repo.update(object_id, data)
-
     def delete(self, object_id):
         return self.repo.delete(object_id)
